backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
google-genai==1.32.0
rank-bm25==0.2.2
